separate arbitrum signer code into own function
