+++ modules/polling/helpers/getArbitrumPollingContractReadOnly.ts
@@ -1,13 +1,7 @@
 import { Contract } from 'ethers';
-import { getArbitrumGoerliTestnetSdk, getArbitrumOneSdk } from '@dethcrypto/eth-sdk-client';
-import { ArbitrumSdkGenerators } from 'modules/web3/types/contracts';
 import { SupportedNetworks } from 'modules/web3/constants/networks';
 import { getGaslessProvider } from 'modules/web3/helpers/chain';
-
-export const arbitrumSdkGenerators: ArbitrumSdkGenerators = {
-  mainnet: getArbitrumOneSdk,
-  goerli: getArbitrumGoerliTestnetSdk
-};
+import { arbitrumSdkGenerators } from './getArbitrumRelaySigner';
 
 export const getArbitrumPollingContractReadOnly = (network: SupportedNetworks): Contract => {
   const sdkNetwork = network === SupportedNetworks.GOERLIFORK ? SupportedNetworks.GOERLI : network;
