+++ modules/polling/helpers/getRelayerBalance.ts
@@ -1,28 +1,13 @@
-import { getArbitrumGoerliTestnetSdk, getArbitrumOneSdk } from '@dethcrypto/eth-sdk-client';
-import { config } from 'lib/config';
-import { ArbitrumSdkGenerators } from 'modules/web3/types/contracts';
 import { SupportedNetworks } from 'modules/web3/constants/networks';
-import { DefenderRelayProvider, DefenderRelaySigner } from 'defender-relay-client/lib/ethers';
 import { formatEther } from 'ethers/lib/utils';
+import { getArbitrumRelaySigner } from './getArbitrumRelaySigner';
 
-export const relayerCredentials = {
-  mainnet: { apiKey: config.DEFENDER_API_KEY_MAINNET, apiSecret: config.DEFENDER_API_SECRET_MAINNET },
-  goerli: { apiKey: config.DEFENDER_API_KEY_TESTNET, apiSecret: config.DEFENDER_API_SECRET_TESTNET }
-};
-
-export const arbitrumSdkGenerators: ArbitrumSdkGenerators = {
-  mainnet: getArbitrumOneSdk,
-  goerli: getArbitrumGoerliTestnetSdk
-};
-
-//Note that we'll get an error if we try to run this defender relay code on the frontend
-//So we should only import this function on the backend
 export const getRelayerBalance = async (network: SupportedNetworks): Promise<string> => {
   const sdkNetwork = network === SupportedNetworks.GOERLIFORK ? SupportedNetworks.GOERLI : network;
-  const provider = new DefenderRelayProvider(relayerCredentials[sdkNetwork]);
-  const signer = new DefenderRelaySigner(relayerCredentials[sdkNetwork], provider);
+
+  const signer = getArbitrumRelaySigner(sdkNetwork);
   const address = await signer.getAddress();
-  const balance = await provider.getBalance(address);
+  const balance = await signer.provider.getBalance(address);
 
   return formatEther(balance);
 };
